Add <noinclude> tags to new <templatedata> strings
If the template is not in a subpage, all new <templatedata> strings
will be wrapped with <noinclude>.

This will not happen to existing <templatedata> strings, since there
is no way of knowing if existing strings are already inside a
<noinclude> chunk - especially if those tags also include a bit of
other text. This will be applied for all new templatedata tags that
are inserted in the main template page.

Bug: 69149
Change-Id: I083a76858513760910973cd788687822fda0e879