Trim aliases on save
Make sure that aliases get trimmed and are not getting extra spaces
when saving a parameter.
NOTE: To prevent dirty diffs, only parameters that are 'touched'
(viewed or edited directly in the editor) will be trimmed. The editor
will keep all other "wrong" values that have trailing or preceding
whitespaces as-is.

Bug: T85330
Change-Id: I64ecb88a7d736fa36b5e6debf2e94b0d37152020