Fix styles queue violation for "ext.templateData"
Only contains a stylesheet and has a dependency. Module is only
loaded via addModuleStyles, which means the dependency is ignored.

It triggers the warning because it has a dependency, which means
it's not a pure styles-only module (despite being used as such)

> Unexpected general module "ext.templateData" in styles queue.

Aside from the addModuleStyles() load, it is also specified as
a dependency for "ext.templateDataGenerator.ui".

Dependencies on style modules from JS modules are supported since
T87871 was resolved, however the indirect dependency still doens't
make sense.

Move the dependency on ext.templateData.images, which wasn't needed
or triggered by "ext.templateData" styles anyway, to the UI module
directly.

This also fixes the fact that the "ext.templateData" css is loaded
twice (once as blocking stylesheet without dependency, and then
again as async module via "ext.templateDataGenerator.ui" with dependency).

Bug: T92459
Change-Id: I4c3bb44c135fccb85a26ad901ec9660d647dfe86