Change "Template preferred format" to "Suggested wikitext formatting"
This sounds clearer to me. The word "preferred" is redundant, because
this is already a preference, but it's replaced with "suggested"
to indicate that it won't necessarily be formatted that way

And "wikitext" should be mentioned explicitly to make clear that it's
about the code, and not about how the template is rendered.

Change-Id: I38dfa9bc406822f14b913d0693a7a2a440586112