Send title without namespace
These are a mess and hard to de-localize.  Also, the schema is
already documented as not including the namespace in this field.

A proper migration would be very difficult, so we'll have to remember
that our data includes this glitch and manually strip namespace
prefixes before the deployment date, as needed.

Change-Id: If2a4dd865b95458dc63162460f252500fd52436e