Use "Edit" for the edit button
We probably used "edit" as strictly speaking it could be "edit"
or "create", and it's hard to tell which will happen before
inspecting the content.

However "manage" is a poor alternative and not a verb we use
anywhere else. "Edit" argualby covers "create" as well, and will
be much more familiar to users.

Bug: T317468
Change-Id: Ia0b5da50fc87bd250e85faf29613a6f97ca36459