Move var declarations (and initializations) down in Dialog class
Note the code style in this file is already a mixture of some early
declarations and others that happen in place. We mostly prefer the
later nowadays because it makes the code much more readable. Some of
the code touched in this patch was really a little confusing, I would
argue.

Change-Id: Ib2777c6f6c86876c016e50d7bd2fe3d82c611914