Add missing dependency on mediawiki.user
The Metrics.js script needs this, but the dependency is not specified
anywhere. This "accidentially" works because it's rather common that
the module is required by some other code on the page. I just found a
combination where this is not the case (a blank MediaWiki installation
with nothing but TemplateData, not even the WikiEditor).

Change-Id: I97a7c1d227316a222d7ca54adb400aa8dd769047