Create dialog elements from "type", not from element name
The data structure in Model.js, starting at line #125, does have a
"type" property that's supposed to do this. What kind of input element
we need depends on the type. Repeating individual element names here
means we are duplicating parts of the data structure.

The mapping is unambiguous, i.e. the new code is guaranteed to do the
same as before.

This is a tiny cleanup patch in preparation for later changes.

Change-Id: I162dcd6f9a31f5ab2c6eb7027befd4736de40ca5