Fix placeholder size for the "edit template data" button
This placeholder was added in Ic8f9bf7. It makes sure the page
content below the button doesn't shift down the moment the button
appears. See T279869 for a screencast that demonstrates the issue.

Unfortunately something changed (possibly related to T323773) and
the size of the placeholder is not correct any more. At the moment
the page shifts up a few pixels.

Bug: T279869
Change-Id: Ie6e585d2061342c11aabb4414bb9b0bedecf3a12