Fix broken ActionFieldLayout construction
OO.ui.ActionFieldLayout requires 3 parameters: 2 widgets and optional
configuration. But this code was passing only 1 widget. The config
object was actually used as a widget and never as config. It's a
miracle this was not throwing an error. ;-)

Replace with a more trivial FieldLayout which requires only 1 widget.

Also clean up some closely related code.

Change-Id: Idca9a39e9f442f3a5412027906de9f868fc12b76