Remove bogus updateParamDetailsLanguage call with no language
Turns out this was always pointless. Nothing the method does makes
sense without a language. A default language is only choosen much
later as part of getSetupProcess, and the method called a second
time, this time with a language.

Issue introduced a long, long time ago and finally surfaced via
Ifcbb877.

Bug: T238329
Change-Id: I47c8ce16107ebe7a1d955665dc890bef80ec892a