Add lang and dir attributes to language search result rows
The languages in this list are presented as "autonyms", meaning it's
the name of the language written in the language. They should be
marked accordingly in the DOM.

However, the current design breaks appart when we allow some of
the rows to be right-aligned. To avoid this we enforce the previous
left-align via CSS. This is not perfect but the best compromise at
the moment, in my opinion.

The code follows what we did in T238329, specifically the patch
Ifcbb877.

Bug: T324385
Change-Id: I53a5a96395b60f3e5d0d80bd5efa4024c4d97884