Remove namespace check
The namespace check is not necessary here since the javascript UI is always
available if a page is in a template data namespace or contains template data.
Therefore this check can be removed.
This change results in the edit links being displayed too on pages which are in
non-templatedata namespaces but contain template data.
Follow-up to 1093972.

Change-Id: I23abbf4bac450857db6a566ede35a3431401cbd2