Category Browser: Add template info display
Do not emit a 'choose' event when a template ColumnItem is clicked,
but instead show an extra column with the template's info, and a
button that emits the event.

This renames the old 'choose' event to be 'select', which is used
when a column item is selected. The 'choose' event from the
button then gets passed up and remains the only external event
that is be exposed by ColumnGroup.

Also adds a list of a template's categories (excluding the one
that's currently selected as the columnItem).

Bug: T395523
Change-Id: I2a08e9bcdb9b57b99c5c8821c584b4fd7bb621c9