refactor: make use of version in schema for validation
As of now, this makes no difference to TemplateData, because the
Schema is still at its initial version 1.0.0. But if in the future the
Schema is modified in a way that requires a migration, then this change
will ensure that wikis will continue working even if they still have a
config for the old version stored on a wikipage, but the new version has
already been deployed via backport or the train.

Bug: T371028
Depends-On: I4d841d03042ce0cdc12384bad97d586145a8ba28
Change-Id: I2f10c5f09e3f0e104947df87554a5f3c3501351c