Remove missing favorites when loading details
When retrieving favorites' details, if any templates are
missing or have been moved to a non-template namespace,
remove them from the favorites array.

This doesn't actually re-save the preference data, but does keep
the displayed info (total favourite count etc.) up to date. The
data will be saved (with the missing templates not inlcuded) the
next time the user faves or unfaves something.

Bug: T388555
Change-Id: I590aaa44a1bbb4e44c09104a1a882c9018dc0b5c