Save FavoritesStore.favoritesArray before saving user option
Save the runtime list of favorites before saving the same value
to the API, in order that multiple calls to saveFavoritesArray
in quick succession will always work with the most up-to-date
value.

This means that if saving fails we're left with an invalid
favoritesArray, so to avoid that we keep the old array and roll
back to it on error. This does mean that a transient error
could possible mean that the values get out of sync.

Bug: T401927
Change-Id: I57c5956433b43213b5bebc14ab9e0fcf738b0458